Add validator keys tool source code (RIPD-1310):
This ripple-libpp based validator key management tool is intended to
replace the Python manifest tool in the rippled git repository.
Validator master keys are generated and stored in a JSON file (default:
$HOME/.ripple/validator-keys.json).
Ephemeral signing keys and corresponding manifest are generated
together using the key file. Manifest sequences are tracked and
automatically incremented in the key file.